Fixed header files after all fields have been stored away cleanly. Hibike write functionality written, and code compiles

+++ hibike/devices/YogiBear/current_limit.cpp
@@ -34,7 +34,7 @@ int limit_state = 0; //tells us which state the FSM is in and how we are modifyi
 
 
 void current_limiting() {
-  double targetPWM;
+  float targetPWM;
   
   if (driveMode == 0) {
     targetPWM = pwmInput;
